requirements.txt
Django==5.2.11
djangorestframework
psycopg2-binary
pillow
gunicorn
